moving Cache-related compiler pass from FrameworkBundle to Cache component
