+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -18,7 +18,7 @@
     "require": {
         "php": "^7.1.3",
         "ext-xml": "*",
-        "symfony/cache": "~3.4|~4.0",
+        "symfony/cache": "~4.2",
         "symfony/dependency-injection": "^4.2",
         "symfony/config": "~4.2",
         "symfony/event-dispatcher": "^4.1",
